refactored ui components
