--- src/tw-styled/ui/formField/FormField.stories.ts
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from "@storybook/react";
-import { FormFieldDemo } from "./FormField.demo";
-
-const meta = {
-  title: "Form Field",
-  component: FormFieldDemo
-} satisfies Meta<typeof FormFieldDemo>;
-export default meta;
-
-export const Demo: StoryObj<typeof meta> = {
-  argTypes: {
-    errorMessage: {
-      control: "boolean"
-    }
-  }
-};
