--- src/tw-styled/ui/hooks.ts
@@ -1,11 +1,0 @@
-import React from "react";
-import { StyleToClassProps } from "../types";
-
-export const useDefaultStyles = (
-  style?: StyleToClassProps,
-  cb?: (style?: StyleToClassProps) => StyleToClassProps
-) => {
-  return React.useMemo(() => {
-    return cb?.(style) || { ...style };
-  }, [style, cb]);
-};
