--- src/tw-styled/ui/index.ts
@@ -1,5 +1,0 @@
-export * from "./fieldError/FieldError";
-export * from "./fieldError2/FieldError";
-export * from "./formField/FormField";
-export * from "./input/Input";
-export * from "./label/Label";
