+++ src/ui/components/fieldError/FieldError.types.ts
@@ -1,0 +1,8 @@
+import { BaseStyles } from "@/tw-styled/types";
+
+export type FieldErrorProps = {
+  message?: string;
+  described_by?: string;
+  error_hidden?: boolean;
+  style?: BaseStyles;
+};
