+++ src/ui/components/formField/FormField.types.ts
@@ -1,9 +1,5 @@
-import { useStyleToClass } from "@/tw-styled/tools";
 import { BaseStyles } from "@/tw-styled/types";
 import { FieldValues, UseFormRegister } from "react-hook-form";
-import { defaultStyles } from "./FormField.defaultStyles";
-import { useFormFieldUi } from "./FormField.ui";
-import { useDefaultStyles } from "../hooks";
 
 export type FormFieldProps<T extends FieldValues> = {
   inputType?: "text" | "password";
@@ -25,14 +21,4 @@ export type FormFieldProps<T extends FieldValues> = {
     errorIcon?: BaseStyles;
     errorInputWrapper?: BaseStyles;
   };
-};
-
-export const FormField = <T extends FieldValues>(props: FormFieldProps<T>) => {
-  const { style, ...rest } = props;
-
-  const styles = useDefaultStyles(style, defaultStyles);
-  const classes = useStyleToClass(styles);
-  const FormField = useFormFieldUi({ classes, ...rest });
-
-  return FormField;
 };
