+++ src/ui/components/index.ts
@@ -1,8 +1,12 @@
 export * from "./box/Box";
 export * from "./button/Button";
+export * from "./fieldError/FieldError";
 export * from "./form/Form";
+export * from "./formField/FormField";
 export * from "./heading/Heading";
 export * from "./icon/Icon";
+export * from "./input/Input";
+export * from "./label/Label";
 export * from "./layout/Layout";
 export * from "./navbar/Navbar";
 export * from "./pulse/Pulse";
