+++ src/ui/components/label/Label.types.ts
@@ -1,0 +1,10 @@
+import { BaseStyles } from "@/tw-styled/types";
+
+export type LabelProps = {
+  options?: {
+    text: string;
+    htmlFor: string;
+    visuallyHidden?: boolean;
+  };
+  style?: BaseStyles;
+};
